Put the summariser's tests where the summariser is
Three of them grew in the dead-letter file because that is where the review
that produced them was looking, and they test `providerFailure` — so someone
opening the summariser's test file to find them would have found one of four.
The row-level PII check stays where it is: that one is about the trail.

Also puts the rule's header back on the function it describes. The last commit
inserted `httpStatus` between them, leaving thirty lines about what a refusal
may say attached to a four-line helper.
