+++ tests/modules/memory-summarize.test.ts
@@ -20,6 +20,7 @@ import {
 import { DATA_FENCE, renderNudge } from "@/graph/nudge";
 import {
   ATTENDANCE_SUMMARY_MAX,
+  providerFailure,
   renderTranscript,
   summarizeAttendance,
 } from "@/modules/memory/summarize";
@@ -314,6 +315,119 @@ describe("summarizeAttendance", () => {
     expect(seen[0]).not.toBe(seen[1]);
     expect(seen[0]?.aborted).toBe(false);
     expect(seen[1]?.aborted).toBe(false);
+  });
+
+  // NOTHING THE SERVER AUTHORED reaches the line — which is a stronger rule than "no prose", and the
+  // weaker one is what an earlier revision shipped. `code` and `type` are vendor error identifiers by
+  // convention only; the value is chosen by the server, this product accepts an arbitrary
+  // OpenAI-compatible endpoint, and a bare token is exactly the shape of a phone number, a CPF or a
+  // first name. So the fields are gone, not filtered.
+  test("nothing the provider authored reaches the line, however clean it looks", () => {
+    const marker = "carambola-com-manjericao-8812";
+    // A single bare token in `code`: no whitespace, no prose, and it would have passed a shape test.
+    const tokenised = providerFailure(
+      Object.assign(new Error("rejected"), {
+        name: "BadRequestError",
+        status: 400,
+        code: marker,
+        type: "invalid_request_error",
+      }),
+    );
+    expect(tokenised).not.toContain(marker);
+    expect(tokenised).not.toContain("invalid_request_error");
+    expect(tokenised).toBe("HTTP 400");
+
+    // The status is read from the client's NUMBER field and nowhere else. Digging it out of the text
+    // was an earlier revision, and the digits were never the point: a 4xx-shaped number in a message
+    // that echoes the transcript is the customer's PIN or their invoice total far more often than it is
+    // a transport status, and naming a status the provider never returned sends the operator to the
+    // wrong thing to fix.
+    const rethrown = providerFailure(
+      new Error(`Request failed with status 429 while processing "${marker}"`),
+    );
+    expect(rethrown).not.toContain(marker);
+    expect(rethrown).toBe("provider error");
+
+    // `name` reads like the SDK's class and is a plain writable property, so a wrapper can assign a
+    // transcript-derived token to it — and a BARE one is exactly what would have survived a shape test.
+    // The field is not read at all now, which is the same answer `code` and `type` got.
+    const wrapped = providerFailure(
+      Object.assign(new Error("boom"), { name: marker, status: 500 }),
+    );
+    expect(wrapped).not.toContain(marker);
+    expect(wrapped).toBe("HTTP 500");
+
+    // `status` is admissible because the client PARSED it into a number, and a number cannot carry a
+    // transcript — so the type check is the whole of the guarantee, not a tidiness. It is not
+    // hypothetical either: Google's error body puts a string in `status` (`INVALID_ARGUMENT`), so a
+    // wrapper copying that field across lands a server-authored string in it.
+    const stringStatus = providerFailure(
+      Object.assign(new Error("boom"), { status: `REJECTED_${marker}` }),
+    );
+    expect(stringStatus).not.toContain(marker);
+    expect(stringStatus).toBe("provider error");
+    // Both spellings go through the one check, so neither is the one that gets it wrong.
+    expect(
+      providerFailure(
+        Object.assign(new Error("boom"), { statusCode: `REJECTED_${marker}` }),
+      ),
+    ).toBe("provider error");
+    expect(
+      providerFailure(Object.assign(new Error("boom"), { statusCode: 503 })),
+    ).toBe("HTTP 503");
+
+    // A number is admissible because it cannot carry a transcript — which covers a number that IS a
+    // status and nothing else. `HTTP NaN` was never in the vocabulary this promises, and 0 (never
+    // connected) and a figure lifted out of the body are not statuses either.
+    // 429.5 is the one that isolates the integer check: every other value here is already refused by
+    // the range, so without it the list passes and `HTTP 429.5` ships.
+    for (const notAStatus of [0, Number.NaN, 429.5, 3.7, 4500, -1, 99]) {
+      expect(
+        providerFailure(
+          Object.assign(new Error("boom"), { status: notAStatus }),
+        ),
+      ).toBe("provider error");
+    }
+    expect(
+      providerFailure(Object.assign(new Error("boom"), { status: 100 })),
+    ).toBe("HTTP 100");
+    expect(
+      providerFailure(Object.assign(new Error("boom"), { status: 599 })),
+    ).toBe("HTTP 599");
+
+    // With nothing to go on, a fixed literal rather than whatever the error happened to be called.
+    const opaque = providerFailure(
+      Object.assign(new Error(marker), { name: marker }),
+    );
+    expect(opaque).toBe("provider error");
+  });
+
+  // The one reading of "it timed out" that the other side does not write. `AbortSignal.timeout` rejects
+  // with a DOMException whose name is "TimeoutError" — a tell living in the same writable field the
+  // rule above stopped trusting — so the signal itself is what decides, and the summariser holds it.
+  test("a summariser that ran out of time says so, from our own signal", () => {
+    const marker = "carambola-com-manjericao-8812";
+    const controller = new AbortController();
+    controller.abort();
+    expect(
+      providerFailure(
+        new Error(`aborted while sending ${marker}`),
+        controller.signal.aborted,
+      ),
+    ).toBe("timeout");
+  });
+
+  // The WIRING of the line above, which no cheap test can drive: making the real timeout fire costs
+  // sixty seconds, and shortening it means a parameter that exists only for the test. So this half is
+  // asserted over the source — and it is worth asserting, because without the argument the summariser
+  // still fails safely and merely reports "provider error" for a timeout, which nothing would notice.
+  // Where the signal is CREATED is not asserted here; that has an observable form, in
+  // tests/modules/memory-summarize.test.ts.
+  test("the summariser decides a timeout from its own signal, not from the error", async () => {
+    const src = await Bun.file("src/modules/memory/summarize.ts").text();
+    expect(src).toContain(
+      "providerFailure(err, attemptSignal?.aborted === true)",
+    );
   });
 
   test("a provider failure is reported, and never throws into the job", async () => {
